Got listenTo events working

+++ db/migrate/20130630003017_create_lists.rb
@@ -2,7 +2,7 @@ class CreateLists < ActiveRecord::Migration
   def change
     create_table :lists do |t|
       t.string :title, null: false
-      t.integer :rank, null: false
+      t.float :rank, null: false
       t.integer :board_id, null: false
 
       t.timestamps
