+++ db/schema.rb
@@ -36,7 +36,7 @@ ActiveRecord::Schema.define(version: 20130708183619) do
   create_table "cards", force: true do |t|
     t.string   "title",       null: false
     t.text     "description"
-    t.integer  "rank",        null: false
+    t.float    "rank",        null: false
     t.datetime "created_at"
     t.datetime "updated_at"
     t.integer  "list_id"
@@ -44,7 +44,7 @@ ActiveRecord::Schema.define(version: 20130708183619) do
 
   create_table "lists", force: true do |t|
     t.string   "title",      null: false
-    t.integer  "rank",       null: false
+    t.float    "rank",       null: false
     t.integer  "board_id",   null: false
     t.datetime "created_at"
     t.datetime "updated_at"
